fix: resolve handicap AttributeError and update architecture docs
- Implement HandicapDecorator as a SQLAlchemy TypeDecorator
- Update mappers, entities, DTOs and unit tests
- Verify 663/663 unit tests and 6/6 E2E tests passing with clean Docker logs
- Ensure correct JSON float serialization for handicaps
- Update CHANGELOG.md to v1.8.0 and CLAUDE.md with architectural decision
- Document when to use TypeDecorator vs composite(), handling nullable value objects and SQLAlchemy VO mapping best practices

+++ src/modules/competition/application/dto/competition_dto.py
@@ -35,7 +35,7 @@ class CreatorDTO(BaseModel):
     first_name: str = Field(..., description="Nombre del creador")
     last_name: str = Field(..., description="Apellido del creador")
     email: str = Field(..., description="Email del creador")
-    handicap: Optional[Decimal] = Field(None, description="Handicap actual del creador")
+    handicap: Optional[float] = Field(None, description="Handicap actual del creador")
     country_code: Optional[str] = Field(None, description="Código ISO del país del creador")
 
     model_config = ConfigDict(from_attributes=True)
